Test fixes and javadocs for attributes and cookies.
