+++ modules/spring-web-test-client/pom.xml
@@ -29,7 +29,7 @@
     <url>http://maven.apache.org</url>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <spring.version>5.0.9.RELEASE</spring.version>
+        <spring.version>5.1.0.RELEASE</spring.version>
         <servlet-api.version>4.0.1</servlet-api.version>
     </properties>
 
@@ -82,13 +82,13 @@
         <dependency>
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-core</artifactId>
-            <version>5.0.8.RELEASE</version>
+            <version>${spring.version}</version>
             <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-test</artifactId>
-            <version>5.0.8.RELEASE</version>
+            <version>${spring.version}</version>
             <optional>true</optional>
         </dependency>
