+++ modules/spring-web-test-client/src/test/java/io/restassured/module/webtestclient/ExtractTest.java
@@ -16,8 +16,8 @@
 
 package io.restassured.module.webtestclient;
 
-import io.restassured.module.webtestclient.http.GreetingController;
 import io.restassured.module.webtestclient.response.WebTestClientResponse;
+import io.restassured.module.webtestclient.setup.GreetingController;
 import org.junit.Test;
 
 import org.springframework.test.web.reactive.server.FluxExchangeResult;
