+++ modules/spring-web-test-client/src/test/java/io/restassured/module/webtestclient/HeaderTest.java
@@ -24,9 +24,9 @@ import io.restassured.function.RestAssuredFunction;
 import io.restassured.http.Header;
 import io.restassured.http.Headers;
 import io.restassured.matcher.ResponseAwareMatcher;
-import io.restassured.module.webtestclient.http.HeaderController;
-import io.restassured.module.webtestclient.http.RedirectController;
 import io.restassured.module.webtestclient.response.WebTestClientResponse;
+import io.restassured.module.webtestclient.setup.HeaderController;
+import io.restassured.module.webtestclient.setup.RedirectController;
 import org.hamcrest.Matcher;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
