+++ modules/spring-web-test-client/src/test/java/io/restassured/module/webtestclient/setup/BasePathController.java
@@ -14,18 +14,32 @@
  * limitations under the License.
  */
 
-package io.restassured.module.webtestclient.http;
+package io.restassured.module.webtestclient.setup;
 
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
 
-public class ParserController {
+/**
+ * This controller has been copied from <a href="http://spring.io/guides/gs/rest-service/">Spring Guides</a>.
+ */
+@Controller
+public class BasePathController {
 
-    @RequestMapping(value = "/parserWithUnknownContentType", method = GET, produces = "some/thing")
-    public @ResponseBody String parserWithUnknownContentType(@RequestParam(value = "param") String param) {
-        return "{ \"param\" : \"" + param + "\" }";
+    private static final String template = "Hello, %s!";
+
+    @RequestMapping(value = "/my-path/greetingPath", method = GET)
+    public @ResponseBody Greeting greeting(
+            @RequestParam(value="name", required=false, defaultValue="World") String name) {
+        return new Greeting(0, String.format(template, name));
+    }
+
+    @RequestMapping(value = "/", method = GET)
+    public @ResponseBody Greeting greeting2(
+            @RequestParam(value="name", required=false, defaultValue="World") String name) {
+        return new Greeting(1, String.format(template, name));
     }
 }
