+++ modules/spring-web-test-client/src/test/java/io/restassured/module/webtestclient/setup/PostAsyncController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.restassured.module.webtestclient.http;
+package io.restassured.module.webtestclient.setup;
 
 import java.util.concurrent.Callable;
