+++ modules/spring-web-test-client/src/test/java/io/restassured/module/webtestclient/setup/ResponseAwareMatcherController.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.restassured.module.webtestclient.http;
+package io.restassured.module.webtestclient.setup;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
